requirements.txt
altair==4.2.0
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
decorator==4.4.2
diffusers==0.2.4
entrypoints==0.4
filelock==3.8.0
ftfy==6.1.1
gitdb==4.0.9
GitPython==3.1.27
huggingface-hub==0.9.0
idna==3.4
imageio==2.22.0
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
MarkupSafe==2.1.1
moviepy==1.0.3
numpy==1.19.5
opencv-python==4.5.5.64
openvino==2022.1.0
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
proglog==0.1.10
protobuf==3.20.2
pyarrow==9.0.0
pydeck==0.8.0b3
pydub==0.25.1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytube==12.1.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
rich==12.5.1
sacremoses==0.0.53
scipy==1.9.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
SpeechRecognition==3.8.1
streamlit==1.12.0
tokenizers==0.13.0
toml==0.10.2
toolz==0.12.0
torch==1.12.1
tornado==6.2
tqdm==4.64.0
transformers==4.16.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
youtube-transcript-api==0.4.4
zipp==3.8.1
